+++ sharding-proxy/src/main/java/io/shardingsphere/proxy/config/RuleRegistry.java
@@ -17,6 +17,7 @@
 
 package io.shardingsphere.proxy.config;
 
+import com.google.common.base.Preconditions;
 import com.google.common.util.concurrent.ListeningExecutorService;
 import com.google.common.util.concurrent.MoreExecutors;
 import io.shardingsphere.core.constant.ShardingProperties;
@@ -25,6 +26,7 @@ import io.shardingsphere.core.constant.TransactionType;
 import io.shardingsphere.core.exception.ShardingException;
 import io.shardingsphere.core.metadata.ShardingMetaData;
 import io.shardingsphere.core.rule.MasterSlaveRule;
+import io.shardingsphere.core.rule.ProxyAuthority;
 import io.shardingsphere.core.rule.ShardingRule;
 import io.shardingsphere.core.yaml.proxy.YamlProxyConfiguration;
 import io.shardingsphere.proxy.metadata.ProxyShardingMetaData;
@@ -48,29 +50,31 @@ import java.util.concurrent.Executors;
  */
 @Getter
 public final class RuleRegistry {
-    
+
     private static final int MAX_EXECUTOR_THREADS = Runtime.getRuntime().availableProcessors() * 2;
-    
+
     private static final RuleRegistry INSTANCE = new RuleRegistry();
-    
+
     private final Map<String, DataSource> dataSourceMap;
-    
+
     private final ShardingRule shardingRule;
-    
+
     private final MasterSlaveRule masterSlaveRule;
-    
+
     private final ShardingMetaData shardingMetaData;
-    
+
     private final boolean isOnlyMasterSlave;
-    
+
     private final ListeningExecutorService executorService = MoreExecutors.listeningDecorator(Executors.newFixedThreadPool(MAX_EXECUTOR_THREADS));
-    
+
     private final String proxyMode;
-    
+
     private final boolean showSQL;
-    
+
     private final TransactionType transactionType;
-    
+
+    private final ProxyAuthority proxyAuthority;
+
     private RuleRegistry() {
         YamlProxyConfiguration yamlProxyConfiguration;
         try {
@@ -91,6 +95,8 @@ public final class RuleRegistry {
         if (!isOnlyMasterSlave) {
             shardingMetaData.init(shardingRule);
         }
+        proxyAuthority = yamlProxyConfiguration.getProxyAuthority();
+        Preconditions.checkNotNull(proxyAuthority.getUsername(), "Invalid configuration for proxyAuthority.");
     }
     
     /**