+++ sharding-proxy/src/main/java/io/shardingsphere/proxy/config/XaProxyRawDataSource.java
@@ -19,7 +19,7 @@ package io.shardingsphere.proxy.config;
 
 import com.atomikos.jdbc.AtomikosDataSourceBean;
 import com.google.common.base.Optional;
-import io.shardingsphere.core.yaml.sharding.DataSourceParameter;
+import io.shardingsphere.core.rule.DataSourceParameter;
 
 import javax.sql.DataSource;
 import java.util.HashMap;